fix: Rename the bin script to mv-j
NPM creates three files in .bin on Windows:

 * <name>
 * <name>.cmd
 * <name>.ps1

If the name contains an extension like .js, the first file will be called name.js
and windows will try executing it instead of name.js.cmd. Names of bin scripts
should not contain suffixes, which might be confused by Windows as file extensions.

BREAKING CHANGE: The name of the executable changed from "mv.js" to "mv-j". I'm sorry
for that, but Windows mistake the suffix ".js" to a file extension and try execute it.
NPM creates the original file name too, probably to support Cygwin.